+++ addons/fray/src/input/sequence/input_requirement.gd
@@ -1,5 +1,5 @@
 extends Resource
-## Used with SequenceData to describe an input sequence.
+## Used by SequencePath to describe an input sequence.
 
 var input: String
 var min_time_held: float