+++ addons/fray/src/input/sequence/sequence_path.gd
@@ -3,11 +3,10 @@ extends Resource
 
 const InputRequirement = preload("input_requirement.gd")
 
-
 ## Array holding the InputRequirements used to detect a sequence.
 var input_requirements: Array # InputRequirement[]
 
-## If true the final input in the sequence is allowed to be triggered with a button release
+## If true the final input in the sequence is allowed to be triggered by a button release
 ## Search 'fighting game negative edge' for more info on the concept
 var allow_negative_edge: bool
 
@@ -23,7 +22,7 @@ func _init(path_allow_nedge = false, inputs: PoolStringArray = []) -> void:
 ##
 ## min_time_held is the minimum time in seconds that the input is required to be held.
 ##
-## Returns a reference to this path to be used in builder-like pattern
+## Returns a reference to this sequence path allowing for chained method calls
 func add(input: String, min_time_held := 0.0, max_delay := .2) -> Reference:
 	var input_requirement := InputRequirement.new()
 	input_requirement.input = input
@@ -34,7 +33,7 @@ func add(input: String, min_time_held := 0.0, max_delay := .2) -> Reference:
 
 ## Setter for 'allow_negative_edge. 
 ##
-## Returns a reference to this path to be used in builder-like pattern
+## Returns a reference to this sequence path allowing for chained method calls
 func enable_negative_edge(allow: bool = true) -> Reference:
 	allow_negative_edge = allow
 	return self